JAVAFIXTURE-27: Improve the package structure (renaming only)
