+++ src/main/java/com/github/nylle/javafixture/SpecimenFactory.java
@@ -1,6 +1,5 @@
 package com.github.nylle.javafixture;
 
-import com.github.nylle.javafixture.generic.FixtureType;
 import com.github.nylle.javafixture.specimen.ArraySpecimen;
 import com.github.nylle.javafixture.specimen.CollectionSpecimen;
 import com.github.nylle.javafixture.specimen.EnumSpecimen;
