+++ src/main/java/com/github/nylle/javafixture/specimen/ArraySpecimen.java
@@ -2,9 +2,9 @@ package com.github.nylle.javafixture.specimen;
 
 import com.github.nylle.javafixture.Context;
 import com.github.nylle.javafixture.CustomizationContext;
+import com.github.nylle.javafixture.FixtureType;
 import com.github.nylle.javafixture.ISpecimen;
 import com.github.nylle.javafixture.SpecimenFactory;
-import com.github.nylle.javafixture.generic.FixtureType;
 
 import java.lang.reflect.Array;
 import java.util.stream.IntStream;
