+++ src/main/java/com/github/nylle/javafixture/specimen/EnumSpecimen.java
@@ -1,8 +1,8 @@
 package com.github.nylle.javafixture.specimen;
 
 import com.github.nylle.javafixture.CustomizationContext;
+import com.github.nylle.javafixture.FixtureType;
 import com.github.nylle.javafixture.ISpecimen;
-import com.github.nylle.javafixture.generic.FixtureType;
 
 import java.util.Random;
