+++ src/main/java/com/github/nylle/javafixture/specimen/PrimitiveSpecimen.java
@@ -1,9 +1,9 @@
 package com.github.nylle.javafixture.specimen;
 
 import com.github.nylle.javafixture.CustomizationContext;
+import com.github.nylle.javafixture.FixtureType;
 import com.github.nylle.javafixture.ISpecimen;
 import com.github.nylle.javafixture.SpecimenException;
-import com.github.nylle.javafixture.generic.FixtureType;
 
 import java.nio.charset.Charset;
 import java.util.Random;
