+++ src/main/java/com/github/nylle/javafixture/specimen/TimeSpecimen.java
@@ -2,9 +2,9 @@ package com.github.nylle.javafixture.specimen;
 
 import com.github.nylle.javafixture.Context;
 import com.github.nylle.javafixture.CustomizationContext;
+import com.github.nylle.javafixture.FixtureType;
 import com.github.nylle.javafixture.ISpecimen;
 import com.github.nylle.javafixture.SpecimenException;
-import com.github.nylle.javafixture.generic.FixtureType;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
