+++ src/test/java/com/github/nylle/javafixture/SpecimenFactoryTest.java
@@ -2,7 +2,6 @@ package com.github.nylle.javafixture;
 
 import com.github.nylle.javafixture.extensions.testcases.TestCase;
 import com.github.nylle.javafixture.extensions.testcases.TestWithCases;
-import com.github.nylle.javafixture.generic.FixtureType;
 import com.github.nylle.javafixture.specimen.ArraySpecimen;
 import com.github.nylle.javafixture.specimen.CollectionSpecimen;
 import com.github.nylle.javafixture.specimen.EnumSpecimen;
