+++ src/test/java/com/github/nylle/javafixture/TypeCaptureTest.java
@@ -1,4 +1,4 @@
-package com.github.nylle.javafixture.generic;
+package com.github.nylle.javafixture;
 
 import org.junit.jupiter.api.Test;
