+++ src/test/java/com/github/nylle/javafixture/specimen/TimeSpecimenTest.java
@@ -2,10 +2,10 @@ package com.github.nylle.javafixture.specimen;
 
 import com.github.nylle.javafixture.Configuration;
 import com.github.nylle.javafixture.Context;
+import com.github.nylle.javafixture.FixtureType;
 import com.github.nylle.javafixture.SpecimenException;
 import com.github.nylle.javafixture.extensions.testcases.TestCase;
 import com.github.nylle.javafixture.extensions.testcases.TestWithCases;
-import com.github.nylle.javafixture.generic.FixtureType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
